+++ src/main/java/com/baldeagle/towny/object/townblock/TownBlockType.java
@@ -1,0 +1,9 @@
+package com.baldeagle.towny.object.townblock;
+
+public enum TownBlockType {
+    NORMAL,
+    SHOP,
+    EMBASSY,
+    ARENA,
+    OUTPOST
+}